+++ pyxel/exposure/__init__.py
@@ -7,6 +7,7 @@
 #
 #
 """Observation classes and function."""
+
 # flake8: noqa
 from .readout import Readout
 from .exposure import Exposure, run_exposure_pipeline